Presenting a basic cleared surface to a Window
This implements the basic infrastructure for sending Vulkan generated
data to a created window.

The associated Surface/Swapchain are automatically resized as the host
window is resized, and presents them in the first provided format and
presentation mode.

The basic RenderPass just clears the screen to a full blue, before that
image is then provided for presentation, with basic inter-frame
semaphores and fences to prevent basic synchronization issues.