Changed Shader Yaml function names
Most complex types have two types of data, much like in programming.

There's the 'declaration' which merely declares a name and perhaps some
basic type information that allows for it to be easily shared many times
lightly.

Then there's the 'definition' which includes all of the details, which
are typically more verbose and only really should be in a single shared
location.