Swapchains now store the extent of the images
On macOS, with the Retina screens, despite requesting perhaps a 1280x720
window, it will instead provide one that is double that size, and yet
the windowing system will report the originally provided size. However,
when generating the Vulkan swapchain, the etent will be provided as the
correct size.

Thus, to deal with this, the swapchain will now store the size of the
associated images, and this value is used for drawing operations to such
surfaces.