Fixed frame indexing
The index would increase even if the next frame wasn't drawn, thus
leading to cases where it would wait on fences that were reset but never
used, and thus would never signal ever again.

This moved the changing of the frame index until after securing the next
frame for rendering to.