Split ImGui state from main application
Along with the split out renderer, the state is now better encapsulated
using an inheritable type that can be embedded into the main menu bar,
or with custom UI elements.

Added a new simple exit option to the file menu, and moved the frame
time info to a separate class.