Fixed vertex descriptor state on pipeline creation
When generating new Vulkan pipelines, using the VertexInput data
directly could be problematic, as the data in the struct could be out
of date. This was because the struct points to other arrays, which as
vectors could be moved around.

To solve, the function `getVertexInputSCI` is used instead, which
updates the struct pointers to wherever the vectors are currently, then
returns a pointer to the required struct.