Split up Vulkan initialization
The single large function dealing with initialization of Vulkan, from
instance to device has been split up, for upcoming new dependencies.

The instance creation/destruction is now separated from devices and
their environments.

This especialls allows for much more flexibility when determining layers
and extensions at both instance and device levels, and for the selection
of the physical device to use.