Changed the abstraction pattern for graphics
Rather than having the full implementation of graphics be in the
foe_graphics library, rather that is being converted to be a set of
common headers for the interface, with the implementation instead being
in an implementation library, such as foe_graphics_vk for Vulkan.

To start this process, the 'device environment' has been replaced with
the runtime and session handle types, with the implementation being
fully abstracted away into implementation-specific libraries.

A bunch of 'temporary' functions are in place for the transition.