Added initial support for Xr isolation
In the main application, we want the foeXrRuntime handle to stay so that
functions don't need different signatures. However, we don't want the
rest of the implementation, as that would being in libraries we don't
have access to.

This uses the compiler definition brought in with the foe_xr_openxr
library to signal what to compile in.