Changed assimp plugin to always be built shared
This model loader is meant to not be statically linked, and always be
pulled in at runtime.

Making the library always shared in CMake ensures it will always be
built as a dynamic library.

As well to go with this, on UNIX platforms the PIC option is now always
used.