Refactored UploadRequest Status checking
Previously, getting the state of an upload request required having
Vulkan-specific headers for both the VkDevice and the return type.

However, with the previous removal of the VkDevice requirement, and
simply creating a new small enum tyoe, the function is now fully abstracted
and moved to the front-end, with no more remaining Vulkan-specific
public functionality.