Implemented EditorNameMap for object names
Names of objects are incredibly useful during development for finding
and referring to items for people as opposed to using plain IDs. However
keeping those names during release simulations are not required, and so
neither would be internalizing them into entity and resource objects.

As such, a separate class will keep them, and ensure the uniqueness of
all of them during development cycles and programs.