foeGfxVkShader now accepts layout creation info
Previously when creating a shader object, it would take in the handle of
the descriptor set layout, requiring it to be acquired beforehand and
passed in. This also added a requirement for a GfxSession function that
performed this operation.

By changing the shader to accept the creation info and perofrming it
internally, we can remove the extra gfxSession function, and make shader
creation just a bit easier for now.