Updated Vulkan struct Yaml parsing script
The parsing of the Vulkan structs is generated using a shell script, and
up until this point would be copied from the generated file for use.

The script has been changed to now also output the header/includes, so
that the whole file can be generated without the copy/paste step.