Updated graphics_vk_yaml export and include dir
For this library, the export was the same as for a possible
graphics_yaml library, which isn't going to end well as it is about to
exist.

As for the include directory, it is currently foe/graphics/yaml, which
again will clash with the new library, requiring the current set to be
shifted to graphics/vk/yaml.