Improved performance enlarging the data pool store
When a larger storage was required previously, everything in the old
store was transferred to the new store before the insertion pass.

This could result in everything in storage being moved up to two times
during the insertion pass.

Now the new storage is created, and an optional pointer is used to point
to it. Then, data is slowly moved, in order, to the new storage so that
*all* data is moved at most once per insert pass.