Split Position3D into separate library
Until now, this type and its associated Yaml functionality has been
situated in the bringup application. However for upcoming Physics items,
it needs to be a separate item that can be brought into that as well.

Thus, while there is very little currently for it, separating it is
necessary. Being a separate library will also keep entanglements lower
and allows for more flexibility with new Position functionality, such as
position chunking.