Resource pools and loaders changed to be derived
The pools are loaders are now derived from common base structs to allow
for the use of inheritance and dynamic casting.

In order to export the vtable and typeinfo on non-MSVC compilers, the
'-fvisibility-ms-compat' compile option has been added. This allows for
the export of these items without having to export the entirety of the
struct or class.