Changed exporting for resource loaders and pools
To avoid exporting whole classes including their members, the use of the
visibility-ms-compat flag on non-MSVCtoolchains was used. However, this
did not work on GCC, so instead falling back to exporting the full
classes, then unexporting the specific members to not be exported.

This will clear up MSVC issues, but other compilers instead offer plenty
or warnings about ignoring the visibility attributes instead.