Added external file searches to importers
Many resources may use external files as a data source during simulation
creation, and as such the ability to search through importers to find
these files is required.

The current loaders that use external files have been changed to search
for the files instead of assuming all paths are based on the current
workingdirectory of the application.