Changed Assimp model importer to not be a plugin
When a loader wants to load data from a model file, the fact that it is
using Assimp should be more explicit rather than relying on flags or
switches.

As such the model plugin system has been removed and anything using it
will now directly use the importer it was explicitly designed for.