Added error codes to simulation initialization
When initializing functionality on any simulation state, there are
plenty of opportunities where failure can occur, and up until now these
failures were hidden.

The initialization of functionality has been modified to properly return
error codes, and where errors do occur, to properly deinitialize based
on the context.

If the functionality is being added new, it then backtracks and removes
the functionality wherever it may have been successfully added and
rejects registering it.

If the initialization fails during the wholesale initialization of a
SimulationState, it then deinitializes that state and returns the code.