Reworked thread pool to possibly fix deadlock
Rather than use a heavy mutex/lock when waiting for all threads to
complete, a cheaper loop is used instead that while yield whenever tasks
haven't yet complete.

This should be much closer behaviour to not using resources than using
just a fast checking loop.

This did require the use of a new atomic variable to bypass possible
data race for the task container, but got rid of a condition variable.