Implemented runner counts for thread pools
While the implementation of the thread pools have become much more
stable, increasing MTTF from 50 to 5000 runs, it is still failing.

The failure seemed to always occur when stopping the pool, after the
threads are waited upon, and whilst the join calls are in progress.

As such, changes have been made to instead keep a count of the number of
threads actually running, and going to join only after they have stopped
running. This also gives the opportunity to notify threads of actions
repeatedly while waiting for them to terminate.