Changed DelayedDestruction abstraction scope
Rather than delayed destruction being at the implementation backend
scope, it has been shifted to the general scope.

This also means that the ability to add functions has been changed to be
generally accessible, with the captured function parameters changed from
VkDevice and VmaAllocator to the generic foeGfxSession handle.

To this end, the implementation has been shifted from the graphics_vk to
graphics library.