Modernized the WSI library
The WSI library was built arounds GLFW3 exclusively, and on the
assumption of only a single window.

With experience since the original implementation, this can be changed
for the better.

WSI now uses the handle mechanism allowing for more than a single
window. As well, the header interface and GLFW3 backend have been split
into separate libraries which should keep things more open and flexible
moving forward.