Isolated exception-handling code and libraries
For the most part, excepptions and their handling is not going to be the
way forward for this project. However some libraries do utilize
exceptions.

So, for thos items, notably dealing with yaml-cpp and CLI11, those
libraries have been isolated and exception left on. For all the others,
exceptions have been turned off.

Code in the engine_lib that use exceptions have beel split off into a
smaller _eh static library for the time being.