Inversed exception handling disabling
Exception handling was previously done by having exceptions being
disabled via compilation options in the root CMake script, then
selectively re-enabling for the libraries with Yaml. However, this also
disabled exceptions in all the test executables which relies on it. It
als added more warnings on MSVC compilations from option overriding.

As such, now excepttion-handling is on by default, and any where it is
disabled is done explicitly on a per-library basis instead.

Also moved adding the source subdirectory to the top of definition
sections.