Replaced GLFW3 functions with universal ones
The GLFW3 function that was used to globally update all windows as the
API allows has been uploaded to the core WSI API as the
foeWsiGlobalProcessing function. In addition, for future APIs that allow
per-window processing the foeWsiWindowProcessing has been added and
implemented.

This also removes the single header file for the GLFW3 implemntation.