Changed foe_xr to an interface library
With the earlier implementation, the idea of how to organize and
properly share export headers was still up in the air. However with the
use of loader libraries how to handle them is now firmed up and the stub
library of foe_xr has been cleared out and fixed.

The foe_xr_openxr implementation now properly exports the header.

The loader for XR has not been implemented yet, as too many OpenXR-
specific details and types are used in the bringup application for a
loader to be effectively used yet.