Changed build behaviour for Window binaries
For Windows to easily link DLLs at runtime, it was best to set a common
binary directory in the build directory.

However, this brought about some confusion when trying to debug static
versions, where these would be sitting in other directories.

To alleviate this, both static and shared builds on Windows will be to
the same directory.