Artificially slow down bringup for ImGui
ImGui is apprently tied to the rate of mouse/keyboard inputs given for
timing. This would lead to issues where mouse/keyboard presses would fly
through the system in less than a millisecond and double-clicks would be
near-impossible, breaking some functionality.

For the moment, an artificial sleep has been added to the main loop to
bring the frame rate down to near 60fps.