Fixed crash with mismatching delayed destructions
When adding a delayed graphics destruction call, if the delay value did
not match the maximal delay, then the provided function list to append
the job to would be invalid.

An extra check after iterating through to check if the iterator is
sitting at the end and fixes it to be at the beginning seems to fix it.