Add ability to set Vulkan session features
When creating a Vulkan device as part of a graphics session, a user can
request features to be enabled via flags through several structs. So the
create function now accepts a void* which will be one of the feature
structs to do so.

As well, passing in similar struct to the new EnumerateSessionFeatures
function allows for querying which features were set on the current
session.