Implemented ImGui render graph job and vk library
The ImGui library has been also dealing with vulkan-specific code. This
goes against what the end goal is supposed to be for modularity, so the
Vulkan-specific renderer class and code has been shifted to the
foe_imgui_vk sub-library.

As well, this new library also holds the render graph job for presenting
the ImGui state in a more simple and self-contained manner.