Add ability to signal fences for present jobs
When presenting a Vulkan graphics image, previously there was no way to
signal a fence, it always had to be done as part of the prior job.

However, now it has been added so a simple submission with no commands
is used after the waiting semaphores are signaled to signal the fence,
then a snew quick semaphore is used to start the image presentation
commands.