Converted OpenXR runtime to use C-arrays
Rather than passing in copies of std::vector types, using the more
portable C-style arrays.

As well, the foeXrRuntime now holds on to the information that it was
created with, including the API version, layers and extensions.

All this information can then be queried later for compatability and
feature support checking.