Added resource state to render graph and jobs
As resources are passed through jobs, currently the state of those
resources is tracked and managed by the callers, both for the input and
output states.

Adding state and dragging it along with resources allows for removing
the need for tracking the ingoing state of resources, slightly
simplifying the use of the render graph.

As well, tracking the state allows for performing more checks at job
creation time, with numerous new specific error codes being added.

Also added were checks to ensure resources are of the correct type,
using the structure chains to confirm the resource types.