Adding graph jobs no longer returns our resources
Previously when adding render graph jobs the function would return the
same number of input resources as output, the only thing it did was to
copy the passed-in job pointer and the input resource data.

Rather, it makes more sense to have the handling of output resources be
handled by the calling context, if it intends to output resources at
all.