Replaced RenderGraphJob with a public handle type
Rather than throwing around raw pointers to the job struct type, instead
the details are provided at the point of adding the job.

The implementation details are kept firmly within the source area of the
render graph, and if data needs to be retrieved later accessor functions
will be created.

Instead, handle types are returned after adding new jobs to the render
graph, and these handles can be used with the resources.