Split graphics initialization for simulations
Until now, the initialization of simulations absolutely required a valid
graphics session to be provided. However, not all simulations will
require presenting graphics or other output. As such, attaching graphics
to these should be optional and separate.

This does require some extra overhead, but the added flexibility will be
more of an asset moving forward.