Remove functions from ResourceLoader base struct
The `canProcessCreateInfo` and `load` functions were determined to be a
required part of the resource loader functionality. Thus, in the
original implementation, these were put in the base struct as pure
virtual functions.

However, as inheritance is being removed, this is no longer going to
work. As such, the functions are being reworked to be passed to the
SimulationState, and would then be called from there instead using
provided function pointers instead.

This also necessitated these functions to be made static, and the loader
function to be split so that the pointer to the loader instance also be
passed in.