Changed sim init selection flag set locations
When simulation functionality is initialized, the location where the
flag is set has been in several different locations. However, it should
be set after the successful incrementing of the related counter.

Before, it used to be set only after the increment and initialization,
or at some other less optimal locations.