Modified foeErrorCode to fix C++ warning
The original intention was to create the type with optional C++
constructors to deal with some fancy implicit behaviours such as
auto-converting from a std::error_code.

However, having these functions as part of the same struct can cause
issues if a compiler decides to add a class pointer which would mess up
the size in the future, leading to undefined behaviour.

Rather, the conversion functions are now separate and explicitly called.