Refactored foeArmature to use new resource pattern
This is a large improvement from the earlier iteration of resources,
where the resource type itself is reduced to a simple struct, no more
code on a resource-specific level.

Iteracting with resources has been reduced to a handle and a dozen
functions, and the complexity and risk of exposing internal workings,
and thus the risk of getting a resource into a bad state have been
massively reduced.

The work for loaders have been reduced where dealing with
synchronization, iteration counts have been eliminated, that has been
offloaded back to the common resource code, allowing loaders to narrow
their focus.