Changed counting of ResourceCreateInfo in Resource
When being brough in via the post-load function, instead of decrementing
it on failure, it is instead being incremented when being copied for
being used as the 'loaded' create info.

As well, the regular createInfo will be decremented/destroyed as part of
the Resource destruction, instead of just doing so for the 'loaded' one.