Made ECS ID header fully C-compatible
The original ECS ID header was only C++ compatible as it also pulled in
and used the limits and strings of the STL.

Most of the header has been moved into the 'id.h' header and contains
all the same data, just derived without the use of the limits library.

The only items not found in this new header are the functions converting
IDs to hexadecimal strings. which have instead been moved to the
'id_to_string.h' C++-only header.