ImGui update for ResourceCreateInfo types
ImGui currently could only support displaying CreateInfo types if it is
the one associated with the same set of registered functionality as the
Resource type itself.

However, CreateInfo types may be from another set, and thus the ability
to find and display arbitrary CreateInfo data from Resource displays was
needed.

To do this, a separate set of functions now need to be registered with
the ImGui registrar.