Implement new foe_graphics_vk_imgui library
This library is used to display graphics_vk types to the UI using ImGui.

Only the specialized VkBool32 type is manually coded. All the other
Vulkan types are auto-generated via the script in the tools directory,
in conjunction with a structs.yaml file that marks which structs to do
and which members of those structs to ignore.