Reworked Yaml exporter responsibility scoping
Yaml exporting is being split along distributed and condensed
capabilities, and to facilitate this there is desired to carve out a
common core set of functionality that generates Yaml node data to be
shared between them.

Unfortunately, the current implementation just directly exported the
data straight to files on an indiviudal basis, so the functionality that
emits that data to files has been 'uploaded' to the calling scopes,
including the details surrounding the filesystem. Instead, these export
clals now deal with returning filled out Yaml data nodes.