Error cases for missing resource/entity importers
When importing data the importers would fail to report any errors
relating to data that could not be processed, or was in fact just
missing. This would lead to issues where initialization would falsely be
considered successful and lead to larger issues later when required data
could not be located.

This changes that, so that if an entity component key cannot be parsed,
then the import process fails immediately, and similarly for resource
data.