Converted registration functions for FFI
Numerous functions scattered about the ImEx and functionality libraries
are used to de/register functionality in a way that is meant to be
possibly used in a portable manner.

However, all these functions were still declared as C++, thus mangled
names and therefore difficult to use from other languages.

To fix, these are put into the extern C namespace, making them to be
formatted as a C FFI.