Added indicator for Yaml importer translator
In some cases an ID translator is not provided, and in such cases it is
not desired to pass it down when retrieving content otherwise it would
give incorrect results.

As the translator data is embedded in the importer object itself, a
boolean flag is used to determine if it has been set or not.