Removed bringup app resource auto-loading
The bringup application has always iterated through all available
resources and explicitly loaded them.

This goes through the various systems and loader that need various
resources to be loaded, and calls for those items to be loaded at these
locations.

Then, because the resources load as requested, removed the explicit
resource load calls at the beginning of the program.