Split Result enum to string to seprate C functions
The current use of foeErrorCode with the std::error_category for getting
string representations of the result codes cannot be made portable
beyond C++ without some undesirable burden of effort.

However, making a set of public functions that take in the code and a
buffer to return the string representations is extremely simple and
portable, inspired by the xrResultToString function found in the OpenXR
library.

As well, the return strings for unknown result values has been made more
succinct, and also returns the unknown number value.

To support negative values, all result enumerations have an added
negative value for the moment.