Move from std::error_code to foeResult
Removed the use of std::error_code throughout the entire codebase,
replaced with the foeResult type. This allows alot of previously
C++-only functions to be FFI compatible.

Removed the use of the vk_error_code mini library. The Vulkan/OpenXR
result to string functions, since they are tiny, are compiled directly
in the libraries that need them from the headers. These headers are
pulled in privately through the khronos_mini_libs interface library.