Refactored foeGfxDelayedDestruction
Changed the declarations to be FFI compatible, and in the process the
calling functions have been changed from using C++'s functional to use
original C-style function pointers with passed-in context.

Renamed to foeGfxDelayedCaller, as while the primary use is for
destruction, it is not the exclusive use and can be used for delaying
other types of graphics calls.