Removed redundant uses of CreateInfo calls
For ResourceCreateInfo, the calls to be able to pull in the data and
subsequently load resource data needed to be public so that they could
be passed into the numerous resource pools. However, with the
unification of the resource data pools into one, it can be made
private to the simulation library.