Add Resource Records for editor work
When in an editing environment, the ability to quicky undo/redo changes
to resources, as well as have historical information is important to
have.

To accomplish this, a new type that tracks these records and has been
added that can sit between the importers/loaders, and the resources
themselves to manage the extra work this entails.