Prevent endless reloading of certain resources
During the main processing loop of the render and physics systems, if a
resource was not specifically in the 'Loaded' state, it would always
request a load.

This did not account for the case where a resource was in the 'Failed'
state, and would endlessly attempt to load the resource regardless of
this state every loop.

To better economize, only attempt to auto-load if it is specifically in
the 'Unloaded' state and not being actively loaded. To re-attempt
loading if a failed resource should be done manually through some other
method.