Convert Gfx Resource CI structs to full C/FFI
The resources for graphics previously had the C++ specific items removed
from them, and the same for the directly related functions.

At this point, the process can be completed where the structs and
functions fully converted to a FFI interface, and the appropriate header
and source files change to be just C files.