Updated ImEx Yaml export process
Sometimes during an save process, an error occurs and the previously
saved data is lost and/or corrupted. This is because the save happens
directly in the desired save location, which is first removed/cleared.

Instead of deleting the desired export directory then writing the
content out there, it now does the export to a temporary location
instead.

Only after all the data could be successfully exported iat the temporary
location will the desired export location be cleared and the data copied.
If the copy fails, an error will occur, but the temporary save in the
temp directory will remain for debugging purposes.