Refined Resource data/create info loading
Can now start loading resource data while create info is being loaded
separately. If this is the case, then the data load will wait for the
create info instead of possibly duplicating the work by yielding the
thread.

If no CI is being loaded and data is requested to load, it will load the
create info itself.