Change type declaration headers for foeRigidBody
One of the members of foeRigidBody is a pointer to a C++ Bullet type.
Because it is a pointer, however, the full type declaration is not
needed at the header, and instead can be included at lowere levels where
the type is used, and as such a forward type declaration has been put in
it's place and the header changed from 'hpp' to 'h'.