Remove resource create info destroy function
In practically all cases, decrementing a ResourceCreateInfo object is
accompanied with a check to see if the reference count reached zero,
then a call to the destroy function within an if block. While this can
work in smaller scale projects, when scaled up the risk increases that
this additional code check will be lost or forgotted, and thus is not
going to work.

Thus the destruction/freeing of ResourceCreateInfo types is now
automatic when the reference count of the object reaches zero through
the ref count decrement function call instead. This does come with the
addition change that the reference count of a created CreateInfo starts
at one instead of zero, but it should be a simple enough concept to
convey and scale much better.