Remove resource destroy function
In the majority of use cases, decrementing a Resource object is
accompanied with a check to see if the reference count hits zero in an
if statement then call the destroy function.

To simplify the interface and avoid issues scaling such code out, the
zero/destroy check has been merged into the ref count decrement call
instead.

As well, the reference count when a resource is created starts at zero
instead of one.

This should make using the Resource type in larger scale projects easier
and simpler.