Add maintenance pass at the end of 'importState'
In the brignup application, the importState call does import all the
content, however it does leave all the component data in the 'insert'
lists of the various component pools instead of the 'regular' section,
which is confusing.

Rather, after the Simulation object is populated, it should be in a
'ready-to-go' state instead.