Removed LogSink base type, use of string_view
The Logger type use to take in the LogSink type to propagate log
messages around, however this was entirely dependent on C++ capabilities
to work.

Instead the logger takes in plain functions and an optional context for
de/registering sinks.

Also changed the log function to use null-terminated char string arrays
instead of a string_view for the message.