Rework thread pool start/stop
Merge the start/stop functionality into the create/destroy calls for
SplitThreadPool.

The start/stop paradigm was originally created when it was a C++ object
based type, where the size of the pools could be easily changed during
runtime. However, such a usecase has not developed, and as such
simplifying the interface is desired.