Changed how swapchain image data is returned
The previous swapchain type would acquire an image, then hold that
image's state until the image was 'released' via a call to the
'presentData' function. This meant however that only one swapchain image
could be operated on at once.

Now, when an image is acquired, all related data is returned at that
time, so now multiple images could be simultaneously in-flight.