Changed how projection/view matrices are uploaded
Rather than having separate systems for 2D views and XR views, the
upload to graphics has been moved to a common RenderViewPool, which
contains a CPU-side simulation buffer to update data possibly multiple
times before updating it in GPU-visible memory.

This also moves the old 'camera' content to be tied with the Window and
XR view data instead of the extra component type, allowing it's
elimination for the time being.