Stop returning errors on advanced Vulkan features
When passing feature structures when creating a Vulkan session, the
session create function would throw an error if a struct was passed in
that belonged to a Vulkan version newer than the one the runtime was
created with.

However, more advanced features can be added on older Vulkan versions
through some layers, such as timeline semaphores. As such, newer feature
structs can still be used on older Vulkan versions at times and should
not be an immediate error.