New logic for Vulkan timeline semaphore
Timeline semaphores are far easier to use and simplify a lot of logic.
As such the enablement is now longer to be tied to XR support, but
should always be added.

However, timeline semaphores can be brought in either via
synchronization2 or a dedicated extension. If neither are found, the
session will fail with an appropriate error code.