Implement simplified ECS ComponentPool type
No longer relying on C++ templates, this version is only compiled once,
with the size of the component type being variable to support different
types.

This also means that the type is more easily usable across different
languages, a much smaller function interface, and the implementation is
tied to whatever is compiled in ECS, opposed to the multiple versions
compiled with each templated variant.