Fix insertPass destructor bug in ComponentPool
When component types that have destructors are pruned out early in the
insert pass, the unique_last call would operate in a manner in which
proper data destruction could not be properly called.

Instead, the filtering out of duplicates will occur during the initial
placement pass, using setting the offset data to INVALID_ID so that the
data is properly destructed in the data copy pass.

Added a function that can destruct an entire data set at once.