Implement EntityList type for ECS
A necessary piece of the ComponentPool is the determination of changes
of held data. Inserts and removes are easy enough to be tracked as part
of the ComponentPool's maintenance call, however data modifications by
external code is not.

The idea is that other code can arbitrarily change component data, and
alert other systems that changes occurred via new EntityLists. These
lists would be added to the appropriate component pool(s) and other
systems could pull these and check for which entities their data had
changed.

This is the initial implementation, along with integration into the
ComponentPool type.