Bringup application fixes
The call to perform window maintenance now captures the result code and
checks it for failure.

When cleaning up swapchains, now checks the swapchain's value for
existance instead of the graphics session.

When checking surface formats and presentation modes, they are stored in
vectors instead of unique_ptr objects for debug readability.