Implement initial render system
The render system deals with taking the RenderState, ArmatureState and
Position3D component data and processing them in a more efficient
manner.

Previously all position and armature data would be processed and put
into memory made available to the renderer, regardless of whether that
data would be used, and when actually processing RenderState, all the
resources would be re-discovered and checked, for every item to be
rendered, every frame.

This system now processes in a smarter fashion, only processing when
component data is inserted/modified/removed, and caching much of it so
that when the render call is performed, it can speed through much faster
than before.