Split runtime armature data to separate component
Currently, both the data saved/loaded externally was sharing the same
ArmatureState structure as the data for the runtime bone matrices.

This is a difficult arrangement to have, especially considering that
would mean having to iterate through the entire set of available
ArmatureState components to determine which need to have per-bone data
calculated each tick.

Rather, these concerns have been split up, so that the ArmatureState
will only deal with base data that can be externalized, and a new
AnimatedBoneState type that deals with the calculated runtime data which
can then be distributed to other systems. This largely reduces the
amount of entities that other systems which could use that data would
need to iterate over, such as for rendering or physics in the future.