Add graphics de/init to VertexDescriptorLoader
This type did not have graphics de/initialization calls since it did not
deal with graphics elements directly. However, this meant that when the
time came to deinitialize, shader-elements still in 'use' by items
loaded by this loader, and thus be unloading in a bad state.

To fix, this type now has de/init calls which are registered to be
called at the right times, so that the vertex shaders are now have usage
counts of zero when unloading.