Pre-increment resources returned by ResourcePool
When a Resource was returned from a ResourcePool, either via the 'add'
or 'find' function, the resource was not incremented.

If a reference count decrement call occurred during the period of time
this was held by another thread, while it was held un-incremented or
before it was incremented by the caller, this could lead to the resource
being destroyed when it should not have been.

Thus, while the ResourcePool lock is still held, the resource to be
returned will be pre-incremented, preventing such an possibility from
occurring.