Remove old importer resource create functions
A byproduct of earlier iterations. Resources used to be loaded at the
same time as the associated create info as part of initializing
simulations, all resources needed to be defined at the start.

This has since changed. Resources are lazy-loaded on-demand, and as such
the creation of resources from their CreateInfo is done as part of the
appropriate selected loader instead.