Remove CreateInfo from resource post-load function
The only purpose left for passing the CreateInfo in was to decrement the
reference count. This meant that the CreateInfo was being held on for
the entire lifetime of the resource load process, even if it was not
needed after a certain point.

Removing it from that call now means the CreateInfo can be decremented
as soon as it is no longer necessary, which can be quite a bit earlier
in the load process.